Issue #22
Removed all scoring attributes except 'score' or 'average_score' from roles and ratings. A rating now has a single score with a value between 1 and 5. A role has one average score. The number of ratings and average score is stored on the role and updated via signals.

Re-organized the javascript. It's organized by application with a commonServices folder. Preparing for JS minification.