Logging and Dependencies
Logging doesn't work as expected with supervisord. Supervisor config changed and requires more work to get python logging to work. However, upervisor does captures stdout without issues, so switching to simple print statements. Dependencies is assumed to be an iterable thing. If it's not, the error is now surfaced appropriately to the API/UI.