workaround 'make dev/tmux_noattach' failing in some cases
For some cases, 'make dev/tmux' and specifically 'make
dev/tmux_noattach' will fail:

        make: *** [dev/tmux_noattach] Error 1

If 'tmux -v -v -v' is used, the error in the logfile is:

        fatal: tty_init: ttyname failed

This seems to be dependant on the version of docker
amongst other things.

https://github.com/moby/moby/issues/8755 is a similar
issue with more detail and background.

But a workaround found there
(https://github.com/moby/moby/issues/8755#issuecomment-83403289)
is to use the 'script' command to avoid the ttyname problems.

This commit uses that workaround for 'make dev/tmux'.