Add 'request_id' and json formatting to galaxy logs
This enables adding or tracking a X-Request-Id per
request that can be used in logging.

https://github.com/dabapps/django-log-request-id

Add 'django-log-request-id' and 'jog' to requirements

Add RequestIdMiddleware to MIDDLEWARE_CLASSES

This adds a log_request_id.filters.RequestIDFilter log filter
to the console/stdout log handlers. The filter adds the
request-id to log records so they can be used by the formatters.

Also add and setup a 'jog.JogFormatter' for formatting normal
python log entries as JSON.