Remove TS unused variables
Currently there are a lot of unused variables in project and they are
mostly represented as unused imports. There are also some unused
variables in scope of module and function definitions as well as some
unused private fields of TS classes.

This patch enables no-unused-variable rule in tslint config.
Due to the fact that it's not possible to check only unused imports,
other rule warnings must be fixed too.
So this patch includes removal of all unused variables found in
function bodies and variables declared but unused in scope of module.

All unused constructor args are removed too since Angular dependency
injection is not dependent on argument order.

Signed-off-by: Ivan Remizov <iremizov@gmail.com>