Use explicit character set instead of "\w" character class
In python 2 "\w" character class refered to ascii set of characters
expressed as "[a-zA-Z0-9_]". In Python 3 it started including
unicode word characters, which is not desirable for expressions
uses in Galaxy.
In Python 3 this behavir can be disabled by using `re.ASCII` flag,
but it is not available in Python 2.
Therefore we replace usage "\w" class with explicit character set
"[a-zA-Z0-9_]" in the project.