Switch to unified logging configuration for galaxy
* Aggregate multiple specific logger configurations
  into single global `galaxy` configuration.
* Set minimal log level `WARNING` for production environment.

Implications:

Due to project structure and celery architecture all worker
messages have to be handled by `galaxy` worker and should not
be propagated to root logger, that is patched by `celery`.
As a result celery log message format is sligtly different from
the galaxy one.
To be fixed in future patches.