Optimize import task list API view
On production database ImportTaskList API call time reaches
6 seconds. The root cause of this problem
is a DISTINCT statement used in SQL query to get all import
task records.

DISTINCT is applied in two places in the code:

1. `GenericAPIView.get_queryset()` method in
   `galaxy/api/views/base_views.py`.
2. `FieldLookupBackend` class in `galaxy/api/filters.py`

To workaround first clause a base class `get_queryset()` method call
is replaced with explicit retrival of model's queryset object.
For a second one a we specify explicitly different filter backend
that doesn't apply `distinct()` to a queryset, because `distinct()`
that is used in `FilterLookupBackend` cannot be removed as it is used
globally in all API views.