Make explicit about how to now do a Local Install (#1326)
Helpful folks on the mailing list, @sivel was able to point me to the contributing guide as a reference to installing locally.  This should have been mentioned in the README.md and a few other notes about the old docker image that the old installer used and is now considered deprecated since 3.0

I've tried to make it super clear about what is going on with these changes.  Corrections welcome if I overspoke anywhere.